Feedback upload refactor
- The feedback upload is proxied through the local HTTP proxy when
  Psiphon is connected.
- Updated psiphon-tunnel-core to 9cb53bb6
