--- Endless/PsiphonTunnel.framework/build-git-commit.txt
@@ -1,1 +1,0 @@
-c56cd6157aa80f391ad208885c83d7a2f6571131
