+++ Endless/PsiphonTunnel.xcframework/ios-x86_64-simulator/PsiphonTunnel.framework/Modules/module.modulemap
@@ -1,0 +1,6 @@
+framework module PsiphonTunnel {
+  umbrella header "PsiphonTunnel.h"
+
+  export *
+  module * { export * }
+}
